requirements.txt
# GFGW-FM: Global Fused Gromov-Wasserstein Flow Matching
# Requirements for one-step high-fidelity image generation

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.0.0

# DINOv2 (installed via torch.hub)
# Feature extraction uses facebookresearch/dinov2

# Optimal Transport
# pot>=0.9.0  # Optional: Python Optimal Transport library

# Metrics
# clean-fid>=0.1.35  # Optional: for FID calculation

# Visualization
matplotlib>=3.5.0
tqdm>=4.62.0

# Configuration
dataclasses-json>=0.5.0

# Distributed training
# accelerate>=0.20.0  # Optional: for easy distributed training

# Logging
wandb>=0.15.0  # Optional: for experiment tracking

# Image loading
pyspng>=0.1.0  # Optional: fast PNG loading
lmdb>=1.4.0    # Required for LSUN dataset

# Development
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
pytest>=7.0.0
